Add comprehensive documentation and improve code quality
- Created a new DOCUMENTATION.md file providing a structured guide for users, developers, and operators, including quick navigation paths and learning paths.
- Updated Makefile to clarify the usage of WITH_LIBPQ variable for PostgreSQL client linkage.
- Enhanced TESTING.md with clearer installation methods and recommendations for users.
- Fixed type issues in ngram.c and pcap_gen.c to ensure proper type casting and prevent potential bugs.
- Cleaned up whitespace in net_config.h and pcap_gen.h for better readability.
- Improved TUI rendering code in tui.c for better clarity and maintainability, including consistent formatting and removal of unnecessary whitespace.
- Updated tui.h to maintain consistent struct definitions and improve code clarity.

+++ src/analysis/ngram.c
@@ -70,7 +70,7 @@ void ngram_train(ngram_model_t *m, const char *text)
     char gram[8];
     gram[m->n] = '\0';
 
-    for (size_t i = 0; i + m->n <= len; i++) {
+    for (size_t i = 0; i + (size_t)m->n <= len; i++) {
         memcpy(gram, text + i, (size_t)m->n);
         ngram_increment(m, gram);
     }
